% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fcts.R
\name{add_obs_string}
\alias{add_obs_string}
\title{Add '.obs' suffix to strings.}
\usage{
add_obs_string(x)
}
\arguments{
\item{x}{String.}
}
\value{
String object with '.obs' suffix.
}
\description{
Function appends strings with '.obs'.
}
