% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_city}
\alias{aggregate_city}
\title{Aggregating wastewater data on city level}
\usage{
aggregate_city(x, pop, measure, merge)
}
\arguments{
\item{x}{Wastewater dataframe.}

\item{pop}{Population dataframe. Df format should be city, location, pop.}

\item{measure}{Gene target of interest for aggregation.}

\item{merge}{Column to merge two dataframes.}
}
\value{
Wastewater dataframe aggregated on city level
}
\description{
Function allows users to aggregate wastewater dataframes on a
 city level for visualization and analysis.
}
