% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fcts.R
\name{build_data}
\alias{build_data}
\title{Create a data object from dataframes.}
\usage{
build_data(cases, hosp, ww, hosp.type)
}
\arguments{
\item{cases}{Dataframe with two columns named \code{date} and \code{value} 
that records the number of cases with respect to time.}

\item{hosp}{Dataframe with two columns named \code{date} and \code{value} 
that records the number of hospitalizations with respect to time.}

\item{ww}{Dataframe with two columns named \code{date} and \code{value} 
that records the viral concentration in wastewater with respect to time.}

\item{hosp.type}{String. Type of hospitalization provided in \code{path.hosp}: \code{NULL}, \code{'hosp.adm'} for hospital admissions, \code{'hosp.occ'} for hospital occupancy.}
}
\value{
A list of dataframes.
}
\description{
User provides dataframes that records the number of cases, (optionally) hospital admissions and viral concentration in wastewater for a specific location.
The various data source do \emph{not} have to have matching date.
}
