% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fcts.R
\name{build_data_csv}
\alias{build_data_csv}
\title{Create a data object from CSV files.}
\usage{
build_data_csv(path.cases, path.hosp, path.ww, hosp.type)
}
\arguments{
\item{path.cases}{String. Path to the CSV file with two columns named \code{date} and \code{value} 
that records the number of cases with respect to time.}

\item{path.hosp}{String. Path to the CSV file with two columns named \code{date} and \code{value} 
that records the number of hospitalizations with respect to time.}

\item{path.ww}{String. Path to the CSV file with two columns named \code{date} and \code{value} 
that records the viral concentration in wastewater with respect to time.}

\item{hosp.type}{String. Type of hospitalization provided in \code{path.hosp}: \code{NULL}, \code{'hosp.adm'} for hospital admissions, \code{'hosp.occ'} for hospital occupancy.}
}
\value{
A list of dataframes.
}
\description{
User provides CSV files that records the number of cases, (optionally) hospital admissions and viral concentration in wastewater for a specific location.
The various data source do \emph{not} have to have matching date.
}
