% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.all}
\alias{calc.all}
\title{Calculate total numbers per epidemiological states 
by summing up values in all compartmental series in input variable comp.
input comp can be E,I,A, and V compartments}
\usage{
calc.all(ts, comp)
}
\description{
Calculate total numbers per epidemiological states 
by summing up values in all compartmental series in input variable comp.
input comp can be E,I,A, and V compartments
}
