% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{calc_concen}
\alias{calc_concen}
\title{calculate average daily deposited RNA concentration}
\usage{
calc_concen(
  ts,
  lambda_I,
  lambda_IH,
  lambda_A,
  lambda_Z,
  mult.shed.t,
  mult.shed.v,
  popSize
)
}
\description{
calculate average daily deposited RNA concentration
}
