% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{calc_delayed_concen}
\alias{calc_delayed_concen}
\title{Estimate RNA/day received at sampling site after decay and delay. 
Using O. Levenspiel, The Dispersion Model, pp. 47-70. 2012.}
\usage{
calc_delayed_concen(ts, kappa, sim.steps, transit.time.mean, transit.time.cv)
}
\arguments{
\item{ts}{Dataframe. Time series from the simulation.}

\item{kappa}{Numeric. Daily decay rate.}

\item{sim.steps}{Integer. Number of time steps per unit of time.}

\item{transit.time.mean}{Numeric. Mean transit time of RNA particles in wastewater from shedding to sampling sites.}

\item{transit.time.cv}{Numeric. Coeeficient ot variation for the transit time of RNA particles in wastewater from shedding to sampling sites.}
}
\description{
Estimate RNA/day received at sampling site after decay and delay. 
Using O. Levenspiel, The Dispersion Model, pp. 47-70. 2012.
}
