% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Rt.R
\name{calc_R_from_model}
\alias{calc_R_from_model}
\title{Estimate Effective Reproduction Number}
\usage{
calc_R_from_model(fitobj, ci, n.cores)
}
\arguments{
\item{fitobj}{List. Results of the fitted model. Output of function \code{fit()}.}

\item{ci}{Numerical. Confidence Interval.}

\item{n.cores}{Numerical. Number of computer CPU cores used for computation.}
}
\value{

}
\description{
Estimate Effective Reproduction Number
}
