% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Rt.R
\name{calc_Reff_from_post}
\alias{calc_Reff_from_post}
\title{Calculate Rt from Posteriors}
\usage{
calc_Reff_from_post(runpost, breaks.time, breaks.val)
}
\arguments{
\item{runpost}{Output of function \code{run_from_posteriors()}}

\item{breaks.time}{Numerical. Time when transmission rate changes.}

\item{breaks.val}{Numerical. Value for the change of transmission rate at
\code{breaks.time}}
}
\value{

}
\description{
This function uses the output of \code{run_from_posteriors()} to
 calculate effective reproductive number.
}
