% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prms.R
\name{check_prior_name_scalar}
\alias{check_prior_name_scalar}
\title{Check if prior's name exist}
\usage{
check_prior_name_scalar(lhs, prm)
}
\arguments{
\item{lhs}{Dataframe of prior samples.}

\item{prm}{List of model parameters.}
}
\value{
Stop the program if at least one inconsistency in the prior parameter names is found.
}
\description{
Checks if the name of all prior parameters are consistent 
with the model parameters already defined. 
The model parameters and the priors parameters are separately defined by the user,
hencethis function checks for typos and inconsistent naming.
}
