% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{define_abc_prms}
\alias{define_abc_prms}
\title{Define parameters for Appoximate Bayesian Computation (ABC) fitting method.}
\usage{
define_abc_prms(iter, accept, case.weight, ww.weight, hosp.weight, hosp.type)
}
\arguments{
\item{iter}{Numeric. Number of prior iterations for ABC fitting.}

\item{accept}{Numeric. Acceptance ratio (so the number of posterior 
samples is \code{iter * accept}).}

\item{case.weight}{Numeric, float. Relative weight for clinical cases}

\item{ww.weight}{Numeric, float. Relative weigth for viral concentration in wastewater.}

\item{hosp.weight}{Numeric, float. Relative weight for hospitalization.}

\item{hosp.type}{String. Type of hospitalization data: 
\code{"hosp.adm"} for hospital admissions, \code{"hosp.occ"} for hospital occupancy and
\code{NULL} for no hospital data.}
}
\value{
Nested list of ABC parameters.
}
\description{
Determine the required parameters (iteration, acceptance and weight)
for the ABC fitting process. 
The weights specifie the importance of different data sources utilized
 in the fitting process simultaneously. There are three data sources 
for simultaneous fitting: reported cases, viral concentration in wastewater 
and hospitalization (admission or occupancy).
}
