% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{define_fit_priors}
\alias{define_fit_priors}
\title{Define the prior distributions for fitting to data.}
\usage{
define_fit_priors(path)
}
\arguments{
\item{path}{String. Path to the CSV file. It need to has a specific format: 
\itemize{
\item column \code{name}: the variable name must match the name of the model parameters that will be fitted. For example \code{R0} (and not, say, \code{Rzero}).
\item column \code{distrib}: the name of the distribution type for the prior variable. Use the standard nomenclature for probability distribution in R (e.g., \code{rnorm()}, \code{rexp()}, ...)
\item column \code{prms}: the parameters for the distribution type for the prior variable. Use the standard nomenclature for probability distribution in R (e.g., \code{mean} and \code{sd} for \code{rnorm()}, etc.) separated by a semi-colon \code{;}. 
}}
}
\value{
Dataframe with three columns: name, distrib,prms.
}
\description{
Create a dataframe with the definition of the prior distribution 
for all parameters that will be fitted to data. The definition are written in a CSV file.
}
