% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{err_fct}
\alias{err_fct}
\title{Calculate error for ABC fitting}
\usage{
err_fct(x, y)
}
\arguments{
\item{x}{Observed data}

\item{y}{Simulated data}
}
\value{
Numerical value for error's fit
}
\description{
Calculate error for ABC fitting
}
