% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Rt.R
\name{estimate_Rt}
\alias{estimate_Rt}
\title{Estimate Effective Reproduction Number.}
\usage{
estimate_Rt(fitobj, ci = 0.95, n.cores = 1)
}
\arguments{
\item{fitobj}{List. Results of the fitted model. Output of function
\code{fit()}.}

\item{ci}{Numerical. Confidence interval (default = 0.95).}

\item{n.cores}{Numerical. Number of computer CPU cores used for computation.
(default = 1)}
}
\value{
A dataframe of estimates for the mean Rt and its confidence interval
 bounds.
}
\description{
This function calculates the effective reproduction number from
 a model fitted on observations. The fitted parameters are sampled from 
 their posterior distribution to simulate epidemic curves. 

 The effective reproduction number, Rt, is estimated from those simulations.
}
