% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst.R
\name{fcst}
\alias{fcst}
\title{Forecast from a fit object.}
\usage{
fcst(fitobj, horizon.fcst, dat, ci = 0.95, n.cores = 1)
}
\arguments{
\item{fitobj}{}

\item{horizon.fcst}{}

\item{dat}{}

\item{ci}{}

\item{n.cores}{}
}
\description{
Forecast from a fit object.
}
