% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fcst_from_post}
\alias{fcst_from_post}
\title{Forecast based on a fitted object.}
\usage{
fcst_from_post(
  path.fitted.object,
  df.ww.cl,
  ci = 0.95,
  time.last.obs = NULL,
  time.horizon = NULL,
  n.cores = 4
)
}
\arguments{
\item{path.fitted.object}{}

\item{df.ww.cl}{}

\item{ci}{}

\item{time.last.obs}{}

\item{time.horizon}{}

\item{n.cores}{}
}
\value{

}
\description{
Forecast based on a fitted object.
}
