% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_col}
\alias{find_col}
\title{find_col(ts, comp) finds columns for a specific compartment
input comp can be E,I,A,H and V and input ts is ODE time-series results
We expect the variable name for compartments to be either
- "comp"  when only there is one compartment: for example "H" 
- "compn" when there are several comaprtments: for example "Hn" and n=1,2,..}
\usage{
find_col(ts, col)
}
\arguments{
\item{ts}{}

\item{col}{}
}
\value{
column's number of compartment col
}
\description{
find_col(ts, comp) finds columns for a specific compartment
input comp can be E,I,A,H and V and input ts is ODE time-series results
We expect the variable name for compartments to be either
- "comp"  when only there is one compartment: for example "H" 
- "compn" when there are several comaprtments: for example "Hn" and n=1,2,..
}
\examples{
find_col(ts,"E")
}
