% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_abc_unit}
\alias{fit_abc_unit}
\title{Perform fitting for one iteration}
\usage{
fit_abc_unit(i, prm.abc, prm, lhs, obs, hosp.var, case.var, initvar)
}
\arguments{
\item{i}{Numerical. The number of iterated simulation for fitting.}

\item{prm.abc}{List. Parameters for ABC fitting.}

\item{prm}{List. All parameters required for model (\code{wem-prm.csv})}

\item{lhs}{Dataframe. Sampled for priors.}

\item{obs}{Datafrmae. Paired clinical-ww data created by \code{build_data()}}

\item{hosp.var}{String. Type of hospitalization data. Output of function \code{build_data()}.}

\item{case.var}{String. Type of date which cases are reported. Output of function \code{build_data()}.}

\item{initvar}{Datafroma defining the initial values for model variables. 
If \code{initvar=NULL}, epidemic starts from a completely naive population.}
}
\value{

}
\description{
Perform fitting for one iteration
}
