% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_recent}
\alias{fit_recent}
\title{Fitting model to recent data only, using existing fit on past data.}
\usage{
fit_recent(
  data,
  prm.abc,
  fitobj.past,
  df.priors,
  prm,
  n.cores,
  last.date = NULL,
  save.rdata = FALSE
)
}
\arguments{
\item{data}{List. Output of function build_data()}

\item{prm.abc}{List. Variables for ABC fitting. Output of function
\code{define_abc_prms()}.}

\item{fitobj.past}{List as returned by the function \code{fit()}. Fit object
for past data.}

\item{df.priors}{Dataframe. Parameters (\code{'name'}), their distribution
(\code{'distrib'}) and range of values (\code{'prms'}) which to be fitted to
the model. Output of function \code{define_fit_priors()}.}

\item{prm}{List. Initial parameters for fitting plot. Output of function
\code{load_prm()}.}

\item{n.cores}{Numeric. number of cores used for fitting computation.}

\item{last.date}{Date. Last date to stop fitting (e.g., ymd('2021-02-01') or
NULL). NULL = latest date available.}

\item{save.rdata}{Logical. Saving the fitting objects in a .rds file.}

\item{do.plot}{Logical. Plot the fitting results and initial parameters}
}
\value{
A list containing the fitted object and additional information.
}
\description{
Fitting model to recent data only, using existing fit on past data.
}
