% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_recent_NEW}
\alias{fit_recent_NEW}
\title{Fitting model to recent data only, using existing fit on past data.}
\usage{
fit_recent_NEW(
  data,
  prm.abc,
  fitobj.past,
  df.priors,
  prm,
  date.pivot,
  n.cores,
  save.rdata = FALSE
)
}
\arguments{
\item{data}{}

\item{prm.abc}{}

\item{fitobj.past}{}

\item{df.priors}{}

\item{prm}{}

\item{date.pivot}{}

\item{n.cores}{}

\item{save.rdata}{}

\item{initvar}{Named numeric vector defining the initial values for model variables.}
}
\value{
A list containing the fitted object and additional information.
}
\description{
Fitting model to recent data only, using existing fit on past data.
}
