% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{generate_obs_noise}
\alias{generate_obs_noise}
\title{Generate observation noise}
\usage{
generate_obs_noise(df, prms = list(cv = 0.1, cv.ww = 0.1))
}
\arguments{
\item{df}{Dataframe. The time series of the epidemiological variables generated
by the function \code{wem::simul(...)$ts}.}

\item{prms}{List of parameters. Must contain elements named \code{cv} and \code{cv.ww} 
to inform the desired coefficient of variation of the added noise for clinical 
reports and concentration in wastewater, respectively.}
}
\value{
The input dataframe with three additional variables, 
\code{report.obs}, \code{hosp.admission.obs} and \code{WWreport.obs}, representing the 
simulated observations of the variables  \code{report}, \code{hosp.admission}  and \code{WWreport}.
}
\description{
Generate observation noise on the reported values for clinical cases
(variable `report`) and concentration in wastewater (variable `WWreport`). 
The observations are generated by drawing samples from a negative binomial distribution
with mean the value from the deterministic model and a variance parametrized
using the coefficient of variation (std. dev / mean).
}
\examples{
prm = model_prm_example() 
sim = simul(prm)
df = generate_obs_noise(sim$ts) 
plot(x=df$time, y=df$report.obs) ; lines(x=df$time, y=df$report)

}
