% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prms.R
\name{load_prm}
\alias{load_prm}
\title{Load all baseline parameters for the wastewater epidemic model.}
\usage{
load_prm(path)
}
\arguments{
\item{path}{String. Path to the model parameters.}
}
\value{
A list of all parameters.
}
\description{
Load all model parameters defined in a CSV file.
Parameters are related to the simulation (population size, time horizon, etc.),
 the epidemic (reproduction number, etc.) and wastewater (decay, delay, etc.) 
The expected format of the CSV file consists of three columns named 
\code{name} for the parameter name, \code{value} for the corresponding value
 and \code{comment} for a brief explanation of the parameter.
}
