% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{merge_fit_past_recent}
\alias{merge_fit_past_recent}
\title{Merge fit object}
\usage{
merge_fit_past_recent(fitobj.past, fitobj.recent)
}
\arguments{
\item{fitobj.past}{}

\item{fitobj.recent}{}
}
\value{

}
\description{
Merge fit object
}
