% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prms.R
\name{model_prm_example}
\alias{model_prm_example}
\title{Example of model parameters}
\usage{
model_prm_example()
}
\value{
List of all model parameters
}
\description{
This is a "helper" function that provides an 
example of the list of all model parameters.
}
