% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{now_string}
\alias{now_string}
\title{Return current date and time with a friendly file name format. Omits year.}
\usage{
now_string()
}
\description{
Return current date and time with a friendly file name format. Omits year.
}
