% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{obs_nbinom}
\alias{obs_nbinom}
\title{Add an observation process distributed as a negative binomial.}
\usage{
obs_nbinom(x, disp)
}
\arguments{
\item{x}{Numerical vector. Mean of observations.}

\item{disp}{Numerical. Dispersion parameter (the larger, the less dispersed)}
}
\value{
A vector the same length as x where each element is a 
random sample with mean x[i].
}
\description{
Add an observation process distributed as a negative binomial.
}
