% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{obs_norm}
\alias{obs_norm}
\title{Add an observation process distributed as normal.}
\usage{
obs_norm(x, cv, do.round = FALSE)
}
\arguments{
\item{x}{Numerical vector. Mean of observations.}

\item{cv}{Numerical. Coefficient of variation.}

\item{do.round}{Logical. Round random samples (default = FALSE).}
}
\value{
A vector the same length as x where each element is a 
random sample with mean x[i].
}
\description{
Add an observation process distributed as normal.
}
