% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overwrite_vectors}
\alias{overwrite_vectors}
\title{Find the names that define an overwrite of vector elements
For example, "v_2" means replace the 2nd element of vector "v"
by the value assigned to prm[["v_2"]].}
\usage{
overwrite_vectors(prm, verbose = FALSE)
}
\arguments{
\item{prm}{List of paramters.}
}
\description{
Find the names that define an overwrite of vector elements
For example, "v_2" means replace the 2nd element of vector "v"
by the value assigned to prm[["v_2"]].
}
