% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_abc_err}
\alias{plot_abc_err}
\title{Visualize Fitting Errors}
\usage{
plot_abc_err(fit, prm.abc)
}
\arguments{
\item{fit}{List. Fitted objects. Output of function fit()}

\item{prm.abc}{List. Required variables for abc fitting.}
}
\description{
Visualize Fitting Errors
}
