% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all}
\alias{plot_all}
\title{Plot all compartments of the model}
\usage{
plot_all(df, ncol = 6, textsize = 12)
}
\arguments{
\item{df}{Dataframe of time series as returned by \code{simul()}.}

\item{ncol}{Integer. Number of column in the faceted plot (default=6).}

\item{textsize}{Integer. Size of the text for the faceted plot (default=12).}
}
\value{
A \code{ggplot2} object displaying a facet plot for all time series.
}
\description{
Plot all compartments of the model
}
