% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dashboard_prm}
\alias{plot_dashboard_prm}
\title{Visualize model parameters}
\usage{
plot_dashboard_prm(prm, nrow = 3, textsize = 10)
}
\arguments{
\item{prm}{List of all model parameters as returned by the function \code{wem::load_prm()}.}

\item{nrow}{Integer. Number of rows in the facetted plot (default = 3)}

\item{textsize}{Intager. Font size for the text in the plot (default = 10)}
}
\value{
A \code{ggplot2} object.
}
\description{
Provide a synthetic view of the model parameter in a dashboard-type format.
}
