% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_epidemic}
\alias{plot_epidemic}
\title{Plot time series of main epidemiological variables.}
\usage{
plot_epidemic(df)
}
\arguments{
\item{df}{Dataframe of time series as returned by \code{simul()}.}
}
\value{
A \code{ggplot2} object displaying a facet plot.
}
\description{
Plot time series of main epidemiological variables.
}
