% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fcst}
\alias{plot_fcst}
\title{Plot forecasts}
\usage{
plot_fcst(var, fcst.obj, dat, date_breaks = "1 month", show.fit = TRUE)
}
\arguments{
\item{var}{}

\item{fcst.obj}{}

\item{dat}{}
}
\description{
Plot forecasts
}
