% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit_abc_vs_obs}
\alias{plot_fit_abc_vs_obs}
\title{Visualize fitting performance versus observational data}
\usage{
plot_fit_abc_vs_obs(prm, post.abc, obs.long, hosp.var, case.var, ci = 0.95)
}
\arguments{
\item{prm}{List. All parameters for model found in \code{wem-prm.csv}.}

\item{post.abc}{List. Contains posterior results. Output of function fit()}

\item{obs.long}{Dataframe. Observational data}

\item{hosp.var}{String. Type of hospital (e.g., \code{NULL},
\code{'hosp.adm'}, \code{'hosp.occ'})}

\item{case.var}{String. Type of date for clinical cases (e.g.,
\code{'report'} and \code{'episode'})}

\item{ci}{Numerical. Percentage of confidence interval.}
}
\value{

}
\description{
Visualize fitting performance versus observational data
}
