% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit_result}
\alias{plot_fit_result}
\title{Visualize Fitting Results and Statistics}
\usage{
plot_fit_result(fitobj, ci = 0.95)
}
\arguments{
\item{fitobj}{List. Fitted object. Output of function \code{fit()}.}

\item{ci}{Numerical. Confidence interval.}
}
\value{
A list of \code{ggplot2} objects:
\itemize{
\item error:  Plot statistical errors from fitting different data sources. 
\item statistic.summary:  Plot statistics of the posteriors. 
\item posterior.dist:  Plot posterior distributions. 
\item fitted.observation:  Plot fitted curve versus observational data. 
}
}
\description{
Visualize Fitting Results and Statistics
}
