% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_multvec}
\alias{plot_multvec}
\title{Plot Transmission Rate (Beta)}
\usage{
plot_multvec(prm, xname, yname, d0, xmax = NULL)
}
\arguments{
\item{prm}{}

\item{xname}{}

\item{yname}{}

\item{d0}{}

\item{xmax}{}
}
\value{

}
\description{
Plot Transmission Rate (Beta)
}
