% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_obs_simulation}
\alias{plot_obs_simulation}
\title{Plot Observational and Simulated Data}
\usage{
plot_obs_simulation(
  df.init,
  obs.long,
  dates.break,
  hosp.var,
  case.var,
  include.cases = TRUE,
  include.ww = TRUE,
  include.hp = TRUE,
  logscale = FALSE
)
}
\arguments{
\item{df.init}{Dataframe. Time-series simulation of all wwmodel's compartments}

\item{obs.long}{Dataframe. Time-series observational data in a long format created by \code{build_obs()}}

\item{dates.break}{Logical. Default is \code{TRUE}. Display break dates when effective reproduction number changes}

\item{hosp.var}{String. Type of hospitalization (e.g., \code{NULL}, \code{'hosp.adm'}, \code{'hos.occ'})}

\item{case.var}{String. Type of date for clinical cases (e.g., \code{'report'} and \code{'episode'})}

\item{include.cases}{Logical.Default is \code{TRUE}. Display simulated and observational reported cases}

\item{include.ww}{Logical. Default is \code{TRUE}. Display simulated and observational measured RNA concentration}

\item{include.hp}{Logical. Default is \code{TRUE}. Display simulated and observational hospitalization}

\item{logscale}{Logical. Default is \code{FALSE}. Logarithmic scale for x-axis}
}
\value{
Visualization for simulation and observational data
}
\description{
For each wastewater sampling location, it plots observational data and corresponding simulation for
reported cases, hospital admissions, hospital occupancy, and measured RNA concentration in wastewater
}
