% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_post_distrib_abc}
\alias{plot_post_distrib_abc}
\title{Visualize posterior distribution}
\usage{
plot_post_distrib_abc(post.abc, df.prior, df.true = NULL)
}
\arguments{
\item{post.abc}{Posterior results. Output of function fit()}

\item{df.prior}{Dataframe. Prior parameters for fitting.}

\item{df.true}{}
}
\value{

}
\description{
Visualize posterior distribution
}
