% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Rt}
\alias{plot_Rt}
\title{Plot the Effective Reproduction Number}
\usage{
plot_Rt(Rdata, min.date, max.date)
}
\arguments{
\item{Rdata}{Dataframe. Rt values from posterior simulation. Output of function \code{estimate_Rt()}}

\item{min.date}{String. Minimum date to show Rt values on the plot.}

\item{max.date}{String. Maximum date to show Rt values on the plot.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the effective reproduction number as calculated by
the function \code{estimate_Rt()}
}
