% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_simobs_beta}
\alias{plot_simobs_beta}
\title{Visualize observational data, simulated epidemic and transmission rate.}
\usage{
plot_simobs_beta(
  data,
  prm,
  include.cases = TRUE,
  include.ww = TRUE,
  include.hosp = FALSE,
  log.scale = FALSE,
  time.range = NULL
)
}
\arguments{
\item{data}{Data object (list) as returned by \code{wem::build_data()}}

\item{prm}{List of all parameter values.}

\item{include.cases}{Logical. Display observations of reported cases?}

\item{include.ww}{Logical. Display observed viral concentration in wastewater?}

\item{include.hosp}{Logical. display hospital observations?}

\item{log.scale}{Logical. Use log scale?}

\item{time.range}{Numerical vector. Minimum and maximum values for the x axis.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the observed data and the simulated epidemic generated 
from the initial parameter values. Also plot the time-varying transmission rate.
This function is typically used to check an initial guess for priors 
(e.g., transmission rate (\code{beta}), wastewater scaling factor (\code{ww.scale})) 
before launching the procedure that fits selected model parameters to data.
}
\examples{
# Load data sets examples from `wem` package:
data('cases')
data('hosp')
data('wwviralconc')

# Build the data object:
dat = build_data(cases = cases, 
                 hosp = hosp, 
                 ww = wwviralconc, 
                 hosp.type = 'hosp.adm', 
                 case.date.type = 'report')

# Load example of model parameters
prm = model_prm_example()

# Plot data, simulation and 
# time-dependent transmission rate (beta)
g = plot_simobs_beta(data          = dat,
                     prm           = prm,
                     include.cases = TRUE,
                     include.ww    = TRUE,
                     include.hosp  = FALSE,
                     log.scale     = FALSE,
                     time.range    = NULL)
plot(g)

}
