% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{post_ss_abc}
\alias{post_ss_abc}
\title{Estimate statistical values after fiiting}
\usage{
post_ss_abc(post.abc, ci.tight = 0.5, ci.broad = 0.95, df.true = NULL)
}
\arguments{
\item{post.abc}{Results of abc fitting.}

\item{ci.tight}{Credible interval}

\item{ci.broad}{Credible interval}

\item{df.true}{}
}
\description{
Estimate statistical values after fiiting
}
