% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prms.R
\name{prm_scalar_vec}
\alias{prm_scalar_vec}
\title{Identify scalar and vector pramaters}
\usage{
prm_scalar_vec(x)
}
\arguments{
\item{x}{dataframe of parameters}
}
\value{
Dataframe with additional columns indicating the type of parameter (scalar vs vector)
}
\description{
Identify scalar and vector pramaters
}
