% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{rebase_prior_names}
\alias{rebase_prior_names}
\title{Rebase priors names}
\usage{
rebase_prior_names(df.priors, fitobj.past)
}
\arguments{
\item{df.priors}{}

\item{fitobj.past}{}
}
\value{
A dataframe with rebased names.
}
\description{
Rebase priors names
}
