% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst.R
\name{remove.fut.inference}
\alias{remove.fut.inference}
\title{Remove future inference in parameters.}
\usage{
remove.fut.inference(idx.t, idx.future, prm, post.abc)
}
\arguments{
\item{idx.t}{}

\item{idx.future}{}

\item{prm}{}

\item{post.abc}{}
}
\description{
Remove elements beyond last date of observed data 
in the vector of time-dependent prms ('variable.t' and 'variable.v') or
in a column priors ('variable_t', 'variable_v')
}
