% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-Rt.R
\name{run_from_posteriors}
\alias{run_from_posteriors}
\title{Run Simulation from Posteriors}
\usage{
run_from_posteriors(fitobj, ci = 0.95, time.horizon = NULL, n.cores = 4)
}
\arguments{
\item{fitobj}{List. Results of the fitted model. Output of function
\code{fit()}.}

\item{ci}{Numerical. Confidence Interval.}

\item{time.horizon}{Numerical. Maximum time for simulation.}

\item{n.cores}{Numerical. Number of computer CPU cores used for computation.}
}
\description{
This function runs a simulation based on the fitted model to
 generate posterior estimates.
}
