% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sample_priors}
\alias{sample_priors}
\title{Randomly sample from priors for fitting process}
\usage{
sample_priors(prior, prm.abc, all.positive = TRUE)
}
\arguments{
\item{prior}{Dataframe. Priors' parameters and distributions. Output of
\code{define_fit_priors()}.}

\item{prm.abc}{List. Varibales required for ABC fitting method created by
function define_abc_prms().}

\item{all.positive}{Logical. Default is \code{TRUE}, which means sampled
values for priors will be truncated to be positive.}
}
\value{
Long dataframe. All the prior sampled values for the parameters to be
 fitted.
}
\description{
This function uses the dataframe output of
 \code{define_priors()} that defines the prior distributions for all
 parameters that must be fitted.
}
