% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{seir}
\alias{seir}
\title{SEIR compartmental model with `nI` I, `nIH` IH, `nA` A, `nH` H and `nZ` Z compartments.}
\usage{
seir(t, x, parms)
}
\arguments{
\item{t}{Numerical. Simulation time.}

\item{x}{List. Initial epidemiological states.   x = (S, Vw, V, Ev1, Ev2, ... nEv,}

\item{parms}{List. Model parameters.}
}
\description{
SEIR compartmental model with `nI` I, `nIH` IH, `nA` A, `nH` H and `nZ` Z compartments.
}
