% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{simul_from_post}
\alias{simul_from_post}
\title{Simulate epidemic from posterior distributions.}
\usage{
simul_from_post(
  post.abc,
  prm,
  hosp.var,
  case.var,
  ci = 0.95,
  n.cores = 4,
  post.slice = NULL
)
}
\arguments{
\item{post.abc}{Dataframe of posterior values.}

\item{prm}{List of (not fitted) model parameters.}

\item{hosp.var}{String. Type of hospitalization; \code{NULL}, \code{'hosp.adm'} and \code{'hosp.occ'}}

\item{case.var}{String. Type of date for clinical cases; \code{'report'} and \code{'episode'}}

\item{ci}{Numeric. Credible interval level. Default = 0.95.}

\item{n.cores}{Integer. Number of cores used for parallel computing.}

\item{post.slice}{EXPERIMENTAL. Default = \code{NULL}.
List with elements specifying the type of posterior
distribution that must be returned.For example, \code{post.slice = list(
t.slice = seq(700, 760, by=10),
vars.slice = c('WWreport', 'inc','clin.case'))}}
}
\value{
Dataframe of summary statistics time series.
}
\description{
Use posterior distributions of a fitted object to simulate an epidemic.
}
