% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{step_line}
\alias{step_line}
\title{Step line function.}
\usage{
step_line(x, b, v)
}
\arguments{
\item{x}{Numeric. Value at which the step line function is evaluated.}

\item{b}{Numeric vector. Breakpoints when the step occurs.}

\item{v}{Numeric vector. Value of the step.}
}
\value{
The y-value where the step function is evaluated
}
\description{
For all x < bb[i], y = vv[i]
}
