% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_abc}
\alias{fit_abc}
\title{Fitting ABC function (internal use)}
\usage{
fit_abc(obs, priors, hosp.var, prm.abc, prm, n.cores = 2)
}
\arguments{
\item{obs}{Dataframe. Paired clinical and ww data. Output of function \code{build_data()}.}

\item{priors}{Dataframe. Sampled priors for fitting. Output of function \code{sample_priors()}}

\item{hosp.var}{String. Output of function \code{build_data()}.}

\item{prm.abc}{List. Parameters for ABC fitting method. Output of function \code{define_abc_prms}}

\item{prm}{List. All parameters model requires. It has to be in the \code{wem-prm.csv} format.}

\item{n.cores}{Numerical. Number of cores for fitting compoutation.}
}
\value{
A list of 1-post fitting variables in a dataframe, 2-priors used for fitting, 3-error of fitting.
}
\description{
Fitting ABC function (internal use)
}
