% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcst.R
\name{remove.fut.inference}
\alias{remove.fut.inference}
\title{Title}
\usage{
remove.fut.inference(idx.t, idx.future, prm, post.abc)
}
\arguments{
\item{idx.t}{}

\item{idx.future}{}

\item{prm}{}

\item{post.abc}{}
}
\value{

}
\description{
Title
}
