% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{simul}
\alias{simul}
\title{Simulate an epidemic}
\usage{
simul(prm)
}
\arguments{
\item{prm}{List of all model parameters.}
}
\value{
A dataframe of the time series of all epidemiological variables.
}
\description{
Simulate an epidemic by solving the system of
ordinary differential equations that define the SEIR-like 
compartmental model.
}
\examples{
prm = model_prm_example() ; sim = simul(prm)

}
