% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{simul_from_post}
\alias{simul_from_post}
\title{Simulate epidemic from posterior distribution}
\usage{
simul_from_post(post.abc, prm, hosp.var, case.var, ci = 0.95, n.cores = 4)
}
\arguments{
\item{post.abc}{Dataframe of posterior values.}

\item{prm}{List of (not fitted) model parameters.}

\item{hosp.var}{String. Type of hospitalization; \code{NULL}, \code{'hosp.adm'} and \code{'hosp.occ'}}

\item{case.var}{String. Type of date for clinical cases; \code{'report'} and \code{'episode'}}

\item{ci}{Numeric. Credible interval.}

\item{n.cores}{Integer. Number of cores for parallel computing}
}
\value{
Dataframe of summary statistics time series.
}
\description{
Simulate epidemic from posterior distribution
}
